ba03134 BUG: use size_t for accumulating image size in numpy conversion

+++ Wrapping/sitkNumpyArrayConversion.cxx
@@ -140,7 +140,7 @@ sitk_GetByteArrayFromImage( PyObject *SWIGUNUSEDPARM(self), PyObject *args )
     size.push_back( sitkImage->GetNumberOfComponentsPerPixel() );
     }
 
-  len = std::accumulate( size.begin(), size.end(), 1u, std::multiplies<size_t>() );
+  len = std::accumulate( size.begin(), size.end(), size_t(1), std::multiplies<size_t>() );
   len *= pixelSize;
 
   // When the string is null, the bytearray is uninitialized but allocated
